refactor pipelines to separate visual, text, and zeroshot. also add pytest skips until model issues are resolved
